requirements.txt
numpy~=1.19.1
pandas~=1.1.1
pyodbc~=4.0.30
PySimpleGUI~=4.29.0

python-dateutil~=2.8.1
REM~=0.2.4
future~=0.18.2
PyYAML~=5.3.1
setuptools~=47.1.0
Jinja2~=2.11.2
pywin32~=228
pdfkit~=0.6.1
pymongo~=3.11.0